starting some main logic
